improve router with additional parameters
create contact page for both technician and householder

routing between missing pages
