+++ src/assets/base.css
@@ -79,8 +79,8 @@ body {
   width: 100%;
   height: 100%;
   z-index: -1;
-  animation: moveGradient 5s ease infinite;
-  background: linear-gradient(-45deg, #a493b4, #aba6ca, #d7d070);
+  animation: moveGradient 10s ease infinite;
+  background: linear-gradient(-45deg, #ceb2fd, #98e9ff, #d7d070);
   background-size: 400% 400%; /* Increase the size for smoother movement */
 }
